requirements.txt
streamlit==1.49.1
Pillow==11.1.0
torch==2.7.1+cu128
torchvision==0.22.1+cu128
fastapi==0.116.0
